ref : 보석 View, ViewModel 나누기
작업 완료
보석
ViewModel 작성

작업 예정
let 뒤에 UI함수를 배치
각각의 UI함수에 viewModel을 달아 data, view 관리

장비
악세서리부분 클릭시 상세설명 나오는 로직 추가
팔찌, 엘릭서 총합, 초월 총합 나오는 로직 추가

보석
보석 Detail 글씨 크기 조절해 볼 생각

카드
이미지 받아서 하는 방식에서 저장된 카드 테두리 사용하는 것으로...
추가 인터페이스

스킬
보석에서 mapping한 정보가 있다면 그것으로 보석 mapping, detail 손보기
